self_agent/requirements.txt
fastapi
pydantic
pydantic-settings
python-jose
langchain-openai
langchain
aiofiles
pytest
